Add lost pins pwm

+++ ESP32_L298N.cpp
@@ -1,7 +1,7 @@
 #include "ESP32_L298N.h"
 
 
-void L298N::moveMotor(int pwmA, int pwmB){
+void L298N::moveMotor(uint16_t pwmA, uint16_t pwmB){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinA,pwmA);
     analogWrite(dl298n.pinB,pwmB);
@@ -15,7 +15,7 @@ void L298N::moveMotor(int pwmA, int pwmB){
     digitalWrite(dl298n.pin2,0);
   }  
 }
-void L298N::moveMotor(int pwmA){
+void L298N::moveMotor(uint16_t pwmA){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinA, pwmA);
     analogWrite(dl298n.pinB, pwmA);
@@ -29,7 +29,7 @@ void L298N::moveMotor(int pwmA){
     digitalWrite(dl298n.pin2,0);
   }    
 }
-void L298N::backMotor(int pwmA, int pwmB){
+void L298N::backMotor(uint16_t pwmA, uint16_t pwmB){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinA,pwmA);
     analogWrite(dl298n.pinB,pwmB);
@@ -43,7 +43,7 @@ void L298N::backMotor(int pwmA, int pwmB){
     digitalWrite(dl298n.pin2,1);
   }
 }
-void L298N::backMotor(int pwmA){
+void L298N::backMotor(uint16_t pwmA){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinA, pwmA);
     analogWrite(dl298n.pinB, pwmA);
@@ -58,17 +58,17 @@ void L298N::backMotor(int pwmA){
   }
         
 }
-void L298N::moveMotor12(int pwmA){
+void L298N::moveMotor12(uint16_t pwmA){
   analogWrite(dl298n.pinA, pwmA);
   digitalWrite(dl298n.pin1,1);
   digitalWrite(dl298n.pin2,0);
 }
-void L298N::backMotor12(int pwmA){
+void L298N::backMotor12(uint16_t pwmA){
   analogWrite(dl298n.pinA, pwmA);
   digitalWrite(dl298n.pin1,0);
   digitalWrite(dl298n.pin2,1);
 }
-void L298N::moveMotor34(int pwmA){
+void L298N::moveMotor34(uint16_t pwmA){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinB, pwmA);
     digitalWrite(dl298n.pin3,1);
@@ -79,7 +79,7 @@ void L298N::moveMotor34(int pwmA){
     digitalWrite(dl298n.pin2,0);
   }
 }
-void L298N::backMotor34(int pwmA){
+void L298N::backMotor34(uint16_t pwmA){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinB, pwmA);
     digitalWrite(dl298n.pin3,0);
